kvserver: move MaxCommandSize to kvserverbase
This commit consolidates the MaxCommandSize cluster setting with its min and
default values. This also helps untangling circular dependencies in the
upcoming changes.

Release note: None
