+++ pkg/kv/bulk/setting.go
@@ -11,7 +11,7 @@
 package bulk
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverbase"
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 )
@@ -30,7 +30,7 @@ var (
 // It returns the smaller of the IngestBatchSize and Raft command size settings.
 func ingestFileSize(st *cluster.Settings) int64 {
 	desiredSize := IngestBatchSize.Get(&st.SV)
-	maxCommandSize := kvserver.MaxCommandSize.Get(&st.SV)
+	maxCommandSize := kvserverbase.MaxCommandSize.Get(&st.SV)
 	if desiredSize > maxCommandSize {
 		return maxCommandSize
 	}
