+++ pkg/kv/kvserver/kvserverbase/base.go
@@ -20,6 +20,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverpb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings"
+	"github.com/cockroachdb/cockroach/pkg/util/humanizeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/quotapool"
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/redact"
@@ -223,4 +224,18 @@ const (
 	// MaxCommandSizeFloor is the minimum allowed value for the
 	// kv.raft.command.max_size cluster setting.
 	MaxCommandSizeFloor = 4 << 20 // 4 MB
+)
+
+// MaxCommandSize wraps "kv.raft.command.max_size".
+var MaxCommandSize = settings.RegisterByteSizeSetting(
+	settings.TenantWritable,
+	"kv.raft.command.max_size",
+	"maximum size of a raft command",
+	MaxCommandSizeDefault,
+	func(size int64) error {
+		if size < MaxCommandSizeFloor {
+			return fmt.Errorf("max_size must be greater than %s", humanizeutil.IBytes(MaxCommandSizeFloor))
+		}
+		return nil
+	},
 )
